+++ gdsfactory/simulation/modes/tests/test_find_modes.py
@@ -11,8 +11,8 @@ def test_find_modes_waveguide():
     # neff1 = 2.3815558509779744
     # neff2 = 1.7749644180250004
 
-    neff1 = 2.3305697149498634
-    neff2 = 1.7019323288562744
+    neff1 = 2.3494603726390664
+    neff2 = 1.7030929743774146
 
     assert np.isclose(m1.neff, neff1), m1.neff
     assert np.isclose(m2.neff, neff2), m2.neff
@@ -21,9 +21,6 @@ def test_find_modes_waveguide():
     modes = find_modes_waveguide(wg_width=0.45, resolution=20)
     m1 = modes[1]
     m2 = modes[2]
-
-    neff1 = 2.3305697149498634
-    neff2 = 1.7019323288562744
 
     assert np.isclose(m1.neff, neff1), m1.neff
     assert np.isclose(m2.neff, neff2), m2.neff